% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{access_data_epa_l4_ecoregions_vsi}
\alias{access_data_epa_l4_ecoregions_vsi}
\title{Access EPA Level IV Ecoregions Data via VSI}
\source{
U.S. EPA Ecoregions Data: \url{https://gaftp.epa.gov/EPADataCommons/ORD/Ecoregions/us/}
}
\usage{
access_data_epa_l4_ecoregions_vsi()
}
\value{
A `sf` (simple features) object containing the EPA Level IV ecoregions shapefile data.
}
\description{
This function retrieves the U.S. EPA Level IV ecoregions shapefile from a remote server via VSI (Virtual Spatial Infrastructure).
The shapefile is stored in a ZIP file, and the function accesses it without downloading the file locally.
}
\details{
The function accesses the EPA Level IV ecoregions shapefile directly from the EPA's data commons, utilizing the `/vsizip/` 
and `/vsicurl/` mechanisms to stream the shapefile from the zipped file. The file is accessed via a URL without the need to 
download it locally. This method allows efficient access to the shapefile data using the `sf` package.
}
\examples{
# Example usage
epa_ecoregions <- access_data_epa_l4_ecoregions_vsi()

}
\references{
U.S. EPA Ecoregions Information: \url{https://www.epa.gov/eco-research/ecoregions-north-america}
}
