% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{access_data_mtbs_conus}
\alias{access_data_mtbs_conus}
\title{Access MTBS CONUS Polygons}
\usage{
access_data_mtbs_conus()
}
\value{
An `sf` object containing the MTBS CONUS polygons.
}
\description{
This function accesses the MTBS (Monitoring Trends in Burn Severity) CONUS (Continental United States) polygons by downloading and reading the MTBS perimeter shapefile directly from the USGS website. The shapefile is accessed via a URL and read into an `sf` object.
}
\examples{
\dontrun{
mtbs_data <- access_data_mtbs_conus()
print(mtbs_data)
}
}
