% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{access_data_welty_jeffries}
\alias{access_data_welty_jeffries}
\title{Access Data from the Welty-Jeffries GIS Server}
\usage{
access_data_welty_jeffries(
  bbox_str,
  epsg_n,
  where_param = "1=1",
  timeout = 600
)
}
\arguments{
\item{bbox_str}{A character string representing the bounding box in the format required by the GIS server. The format is usually "xmin,ymin,xmax,ymax".}

\item{epsg_n}{An integer specifying the EPSG code for the spatial reference system to be used for the input and output data.}

\item{where_param}{A character string representing the SQL WHERE clause to filter the data. Default is "1=1", which retrieves all records.}

\item{timeout}{An integer specifying the timeout (in seconds) for the server request. Default is 600 seconds.}
}
\value{
A `sf` (simple features) object containing the geospatial data retrieved from the Welty-Jeffries GIS server.
}
\description{
This function retrieves geospatial data from the Welty-Jeffries GIS server using the specified bounding box and EPSG code. It constructs a URL query and sends a request to the server to fetch data in GeoJSON format.
}
\examples{
\dontrun{
# Example usage:
bbox <- "-109,36,-102,41"
epsg <- 4326
data <- access_data_welty_jeffries(bbox_str = bbox, epsg_n = epsg)
}
}
