% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{col2hex}
\alias{col2hex}
\title{Convert R Color to Hexadecimal}
\usage{
col2hex(color)
}
\arguments{
\item{color}{A character string specifying a standard R color name.}
}
\value{
A character string representing the hexadecimal color code of the specified R color.
}
\description{
This function converts a standard R color name (e.g., 'red', 'steelblue') to its corresponding hexadecimal color code.
}
\examples{
# Convert the color 'red' to its hexadecimal equivalent
col2hex("red")

# Convert the color 'steelblue' to its hexadecimal equivalent
col2hex("steelblue")

}
