% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_conversions.R
\name{convert_pdf_to_png}
\alias{convert_pdf_to_png}
\title{Convert PDF to PNG}
\usage{
convert_pdf_to_png(pdf_file, dpi, out_dir = NA)
}
\arguments{
\item{pdf_file}{A character string specifying the path to the PDF file to be converted to PNG format.}

\item{dpi}{An integer specifying the required DPI (dots per inch) of the converted PNG.}

\item{out_dir}{A character string specifying the directory to output the converted file. If left empty, the converted file will be saved in the same directory as the original file.}
}
\value{
No return value. The function saves the PNG file to the specified location.
}
\description{
This function converts a PDF file to a PNG image with a specified DPI (dots per inch). It allows the output to be saved in a specified directory or in the same directory as the original PDF file.
}
\examples{
# Convert all PDF files in a directory to PNG format with 300 DPI
pdfs <- list.files(here::here('figs'), pattern = "\\\\.pdf$", full.names = TRUE, ignore.case = TRUE)
pdfs |> purrr::walk(~ convert_pdf_to_png(pdf_file = .x, dpi = 300))

# Convert a single PDF file to PNG format with 300 DPI and save it in a specific directory
convert_pdf_to_png(pdf_file = "example.pdf", dpi = 300, out_dir = "output_directory")

}
