% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_conversions.R
\name{convert_to_cmyk}
\alias{convert_to_cmyk}
\title{Convert Image File to CMYK Format Using ImageMagick}
\usage{
convert_to_cmyk(file, out_dir = NA)
}
\arguments{
\item{file}{A character string specifying the path to the file to convert to CMYK format.}

\item{out_dir}{A character string specifying the directory to output the converted file. If left empty, the converted file will be saved in the same directory as the original file.}
}
\value{
No return value. The function outputs the converted file to the specified directory.
}
\description{
This function converts an image file to CMYK format using ImageMagick. ImageMagick must be installed and accessible via the command line for this function to work.
}
\examples{
# Convert all PNG files in a directory to CMYK format
pngs <- list.files(here::here('figs'), pattern = "\\\\.png$", full.names = TRUE, ignore.case = TRUE)
pngs |> purrr::walk(~ convert_to_cmyk(file = .x))

# Convert a single file to CMYK format and save it in the same directory
convert_to_cmyk(file = "example.png")

}
