% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{dir_ensure}
\alias{dir_ensure}
\title{Ensure Directory Exists}
\usage{
dir_ensure(path)
}
\arguments{
\item{path}{A character string specifying the path to the new directory.}
}
\value{
The function does not return any value. It creates a directory if it does not already exist.
}
\description{
This function checks if a directory exists at the specified path, and if not, creates a new directory.
}
\examples{
# Ensure a directory named "data" exists
dir_ensure("data")

}
