% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{download_data_from_gdrive}
\alias{download_data_from_gdrive}
\title{Download a file from Google Drive to a local directory}
\usage{
download_data_from_gdrive(gDrivePath, localPath)
}
\arguments{
\item{gDrivePath}{A character string. The path or name of the file on Google Drive.}

\item{localPath}{A character string. The local path where the file will be saved.}

\item{overwrite}{A logical value indicating whether to overwrite the file if it already exists at the local path. Defaults to `TRUE`.}
}
\value{
The downloaded file will be saved to the specified `localPath`.
}
\description{
This function downloads a file from a Google Drive path to a specified local path.
}
\details{
This function retrieves a file's ID from Google Drive using the provided `gDrivePath` and downloads it to the local directory specified by `localPath`. The file will be overwritten if `overwrite` is set to `TRUE` (default).
}
\note{
You must be authenticated with Google Drive via the `googledrive` package for this function to work.
}
\examples{
\dontrun{
# Example usage:
download_data_from_gdrive("path/to/file/on/drive", "path/to/local/file.csv")
}

}
