% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{save_kable_workaround}
\alias{save_kable_workaround}
\title{Save Kable Output as PNG with Workaround}
\usage{
save_kable_workaround(k, filePath)
}
\arguments{
\item{k}{An output object from the `kable` function.}

\item{filePath}{A character string specifying the full desired file path (e.g., 'myDir/figs/myTable.png') for the output PNG file.}
}
\value{
No return value. The function saves the PNG file to the specified location.
}
\description{
This function provides a workaround for an issue (as of 3/20/24) with `kableExtra::save_kable`, which fails to export tables as `.png` files. It first saves the table as an HTML file and then converts it to a PNG using `webshot2`.
}
\examples{
# Save a kable output as a PNG file
\dontrun{
k <- knitr::kable(head(mtcars))
save_kable_workaround(k, "myDir/figs/myTable.png")
}

}
