% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographic.R
\name{st_crop_careful_universal}
\alias{st_crop_careful_universal}
\title{Clip a raster to a vector with careful handling of projections}
\usage{
st_crop_careful_universal(raster, vector, mask = FALSE, verbose = FALSE)
}
\arguments{
\item{raster}{A SpatRaster, PackedSpatRaster, RasterLayer, RasterStack, or RasterBrick object. 
The raster to be clipped.}

\item{vector}{A SpatVector, PackedSpatVector, or sf object. The vector defining the clipping boundary.}

\item{mask}{Logical. Should the raster be masked to the vector? Defaults to `FALSE`.}

\item{verbose}{Logical. If `TRUE`, provides detailed output of the steps being performed.}
}
\value{
A raster object clipped to the vector's extent, returned in the same format as the input raster.
}
\description{
This function clips a raster to the extent of a vector, ensuring that the
raster and vector are in the same projection system. It supports raster and
vector objects from both the `terra` and `raster` packages. If the input raster
or vector is packed (for parallelized workflows), it will be unpacked before
processing and optionally re-packed afterward.
}
\examples{
\dontrun{
raster_obj <- terra::rast(system.file("ex/logo.tif", package = "terra"))
vector_obj <- terra::vect(system.file("ex/logo.shp", package = "terra"))
cropped_raster <- st_crop_careful_universal(raster_obj, vector_obj, mask = TRUE, verbose = TRUE)
}
}
