% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{substr_right}
\alias{substr_right}
\title{Substring from the Right}
\usage{
substr_right(str, n)
}
\arguments{
\item{str}{A character string from which to extract the substring.}

\item{n}{An integer specifying the number of characters to keep from the right of the string.}
}
\value{
A character string containing the rightmost \code{n} characters of the input string.
}
\description{
This function extracts a substring from the right side of a given string, retaining a specified number of characters.
}
\examples{
# Extract the last 3 characters from a string
substr_right("Hello, World!", 3)

}
