% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{timestamp}
\alias{timestamp}
\title{Generate a Timestamp}
\usage{
timestamp()
}
\value{
A character string representing the current date and time in the format "YYYY-MM-DD HH:MM:SS TZ".
}
\description{
This function generates a timestamp in the format "YYYY-MM-DD HH:MM:SS TZ", representing the current date and time along with the time zone.
}
\examples{
# Generate a timestamp
timestamp()

}
